requirements.txt
streamlit
python-dotenv
PyPDF2
python-docx
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
